carousel

+++ Scripts/zip-projects.py
@@ -16,11 +16,13 @@ for fileName in os.listdir(path):
 	if not filePath.endswith(".framer"):
 		continue
 
-	# Copy index.html template
-	shutil.copyfile(
-		os.path.join(os.path.dirname(__file__), "..", "Data", "example.html"),
-		os.path.join(filePath, "index.html"),
-	)
+	# Copy index.html template if not exists
+
+	if not os.path.exists(os.path.join(filePath, "index.html")):
+		shutil.copyfile(
+			os.path.join(os.path.dirname(__file__), "..", "Data", "example.html"),
+			os.path.join(filePath, "index.html"),
+		)
 
 	print "Zipping example project '%s'" % filePath
